hotel_booking/requirements/base.txt
pytz==2022.2.1  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.3.4  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.3.0  # https://github.com/celery/django-celery-beat
flower==1.1.0  # https://github.com/mher/flower
uvicorn[standard]==0.18.2  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.2.15  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.51.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.23.1  # https://github.com/tfranzel/drf-spectacular

djangorestframework-simplejwt==4.6.0
pyotp==2.6.0
django-versatileimagefield==2.0
django-phonenumber-field==6.4.0
django-phonenumbers==1.0.1
